Remove `yaml-test-suite` from Cargo releases.
This removes the need to include its license in the Cargo release, where
the tests should not be run anyway. Print a warning still if the folder
isn't there, so contributers can be aware of its absence.

Also removed some explicit mentions of the current version, which makes
it easier to release.